[core] Parquet becomes default file format (#3663)

+++ paimon-flink/paimon-flink-common/src/test/java/org/apache/paimon/flink/action/SortCompactActionForUnawareBucketITCase.java
@@ -139,7 +139,7 @@ public class SortCompactActionForUnawareBucketITCase extends ActionITCaseBase {
                 .createReader(dataSplit)
                 .forEachRemaining(
                         a -> {
-                            Integer current = a.getInt(2);
+                            int current = a.getShort(2);
                             Assertions.assertThat(current).isGreaterThanOrEqualTo(i.get());
                             i.set(current);
                         });
